requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyMySQL==1.0.2
testing.common.database==2.0.3
testing.mysqld==1.4.0
Werkzeug==1.0.1
